修改了3/2=1的bug.

+++ src/main/java/com/github/hcsp/datatype/Main.java
@@ -5,8 +5,8 @@ public class Main {
     // 例如 3/2 == 1.0
     // 请修复此问题，让3/2的结果为正确的1.5
     // 不要修改方法的参数类型
-    public static double divide(double a, double b) {
-        return a / b;
+    public static double divide(int a, int b) {
+        return (double)a / b;
     }
 
     public static void main(String[] args) {
